// src/app/interfaces/all-movies.interface.ts
export interface AllMoviesList {
    rank:        number;
    title:       string;
    thumbnail:   string;
    rating:      string;
    id:          string;
    year:        number;
    image:       string;
    description: string;
    trailer:     string;
    genre:       string[];
    director:    string[];
    writers:     string[];
    imdbid:      string;
}